pyproject.toml
[project]
name = "process-guide-agent"
version = "0.1.0"
description = "ADK agent that guides employees through a company process described in a document"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "python-docx",
    "pdfplumber",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"
